Extract shared frontend helpers
