+++ js/urls-COV-UP-R.chunk.mjs
@@ -1,0 +1,2 @@
+const d=()=>{const n=(window.location.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)(?:\/.*)?$/);return n&&n[1]||null},s=n=>{if(!n||typeof n!="string")return null;let i;try{i=new URL(n,window.location.origin)}catch{return null}const t=decodeURIComponent(i.pathname||"");if(!t.endsWith(".pad"))return null;const e=["/remote.php/dav/files/","/public.php/dav/files/"].find(u=>t.includes(u));if(!e)return null;const o=t.indexOf(e),l=t.substring(o+e.length),r=l.indexOf("/");return r===-1?null:"/"+l.substring(r+1)};export{s as a,d as p};
+//# sourceMappingURL=urls-COV-UP-R.chunk.mjs.map
