+++ js/urls-COV-UP-R.chunk.mjs.license
@@ -1,0 +1,7 @@
+SPDX-License-Identifier: AGPL-3.0-or-later
+SPDX-FileCopyrightText: etherpad-nextcloud developers
+
+This file is generated from multiple sources. Included packages:
+- etherpad-nextcloud
+	- version: 1.0.0
+	- license: AGPL-3.0-or-later
