+++ src/lib/constants.js
@@ -1,0 +1,8 @@
+/**
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ * Copyright (c) 2026 Jacob Bühler
+ */
+
+export const APP_ID = 'etherpad_nextcloud'
+export const MIME = 'application/x-etherpad-nextcloud'
+export const VIEWER_HANDLER_ID = 'etherpad_nextcloud'
